performance hacks
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@5288 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/plasma/parser/mimeType/odtDetector.java
@@ -82,7 +82,7 @@ public class odtDetector implements MagicDetector {
             
             // read in the content of the file
             final InputStream zippedContent = zipFile.getInputStream(mimeTypeInfo); 
-            final String realMimeType = new String(serverFileUtils.read(zippedContent, mimeTypeInfo.getSize()));
+            final String realMimeType = new String(serverFileUtils.read(zippedContent, (int) mimeTypeInfo.getSize()));
             
             return new String[]{realMimeType};
         } catch (final Exception e) {
